Balls can shrink

+++ src/ball.cpp
@@ -16,12 +16,20 @@ void Ball::kill(int ballNum, bool giveScore) {
 
 void Ball::push(int ballNum, Vector2 pushSteps) {
 
-
-    ballPos[ballNum] += pushSteps;
-    if (!Pos::isPosInScreen(ballPos[ballNum])) {
-
-        Ball::kill(ballNum, false);
+    while (Pos::isClippingOutsideScreen(Data::ball, ballPos[ballNum] + pushSteps, ballNum)) {
+        if (!Pos::isPosInScreen(ballPos[ballNum])){
+            break;
+        }
+        if (ballSize[ballNum] <= 10) {
+            Ball::kill(ballNum, false);
+        }
+        ballSize[ballNum]--;
     }
+
+        ballPos[ballNum] += pushSteps;
+
+
+
     if (Utils::haveCollision(Data::player, Data::ball, ballNum)) {
         Ball::kill(ballNum);
     }
